LayoutTests/imported/w3c/web-platform-tests/css/css-color/system-color-consistency-expected.txt


FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a button (light) assert_equals: expected "rgba(0, 0, 0, 0.847) rgb(192, 192, 192)" but got "rgb(255, 255, 255)"
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a button (light)
PASS Property color value 'ButtonText' resolves to the same color as text on a button (light)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a submit button (light) assert_equals: expected "rgba(0, 0, 0, 0.847) rgb(192, 192, 192)" but got "rgb(255, 255, 255)"
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a submit button (light)
PASS Property color value 'ButtonText' resolves to the same color as text on a submit button (light)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a reset button (light) assert_equals: expected "rgba(0, 0, 0, 0.847) rgb(192, 192, 192)" but got "rgb(255, 255, 255)"
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a reset button (light)
PASS Property color value 'ButtonText' resolves to the same color as text on a reset button (light)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a color picker (light) assert_equals: expected "rgb(0, 0, 0) rgb(128, 128, 128)" but got "rgb(255, 255, 255)"
FAIL Property color value 'ButtonFace' resolves to the same color as the background-color of a color picker (light) assert_equals: expected "rgb(255, 255, 255)" but got "rgb(192, 192, 192)"
FAIL Property color value 'ButtonText' resolves to the same color as text on a color picker (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'CanvasText' has the same color as the color of the html element (light)
PASS Property color value 'Field' resolves to the same color as the background-color of a text field (light)
FAIL Property color value 'FieldText' resolves to the same color as text on a text field (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a password field (light)
FAIL Property color value 'FieldText' resolves to the same color as text on a password field (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a email field (light)
FAIL Property color value 'FieldText' resolves to the same color as text on a email field (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a number field (light)
FAIL Property color value 'FieldText' resolves to the same color as text on a number field (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a date field (light)
FAIL Property color value 'FieldText' resolves to the same color as text on a date field (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a text area (light)
FAIL Property color value 'FieldText' resolves to the same color as text on a text area (light) assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Mark' has the same color as the background-color of a mark element (light)
PASS Property color value 'MarkText' has the same color as the color of a mark element (light)
PASS Property color value 'LinkText' has the same color as the color of an anchor element (light)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a button (dark) assert_equals: expected "rgba(255, 255, 255, 0.847) rgba(255, 255, 255, 0.247)" but got "rgb(255, 255, 255)"
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a button (dark)
PASS Property color value 'ButtonText' resolves to the same color as text on a button (dark)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a submit button (dark) assert_equals: expected "rgba(255, 255, 255, 0.847) rgba(255, 255, 255, 0.247)" but got "rgb(255, 255, 255)"
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a submit button (dark)
PASS Property color value 'ButtonText' resolves to the same color as text on a submit button (dark)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a reset button (dark) assert_equals: expected "rgba(255, 255, 255, 0.847) rgba(255, 255, 255, 0.247)" but got "rgb(255, 255, 255)"
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a reset button (dark)
PASS Property color value 'ButtonText' resolves to the same color as text on a reset button (dark)
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a color picker (dark) assert_equals: expected "rgb(255, 255, 255) rgb(128, 128, 128)" but got "rgb(255, 255, 255)"
FAIL Property color value 'ButtonFace' resolves to the same color as the background-color of a color picker (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'ButtonText' resolves to the same color as text on a color picker (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
PASS Property color value 'CanvasText' has the same color as the color of the html element (dark)
FAIL Property color value 'Field' resolves to the same color as the background-color of a text field (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'FieldText' resolves to the same color as text on a text field (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
FAIL Property color value 'Field' resolves to the same color as the background-color of a password field (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'FieldText' resolves to the same color as text on a password field (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
FAIL Property color value 'Field' resolves to the same color as the background-color of a email field (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'FieldText' resolves to the same color as text on a email field (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
FAIL Property color value 'Field' resolves to the same color as the background-color of a number field (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'FieldText' resolves to the same color as text on a number field (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
FAIL Property color value 'Field' resolves to the same color as the background-color of a date field (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'FieldText' resolves to the same color as text on a date field (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
FAIL Property color value 'Field' resolves to the same color as the background-color of a text area (dark) assert_equals: expected "rgb(30, 30, 30)" but got "rgba(255, 255, 255, 0.247)"
FAIL Property color value 'FieldText' resolves to the same color as text on a text area (dark) assert_equals: expected "rgb(255, 255, 255)" but got "rgba(255, 255, 255, 0.847)"
PASS Property color value 'Mark' has the same color as the background-color of a mark element (dark)
PASS Property color value 'MarkText' has the same color as the color of a mark element (dark)
PASS Property color value 'LinkText' has the same color as the color of an anchor element (dark)
